changed int to long, added findbystatus test

+++ PetStoreAPITestingApp/src/test/java/org/example/petstoreapitestingapp/pets/PetsTestBase.java
@@ -6,7 +6,7 @@ import org.example.petstoreapitestingapp.pojo.Pet;
 
 public abstract class PetsTestBase extends TestBase {
     protected static Response response;
-    protected Integer petId;
+    protected Long petId;
     protected static Pet petResponse;
     protected static Pet[] pets;
 }
